you can add product to cart, but choosing quantity doesn't work yet
